# docs/Architecture/openapi.yaml
openapi: 3.1.0
info:
  title: 闪卡 App 后端 API
  version: 2.5.0
  description: >
    闪卡 App v2.5 接口契约(机器可读版)。行为契约(状态机、幂等、FSRS 排程、
    错误码语义)见 structure-contract.md;需求依据 PRD v2.5(增量继承 v2.4~v2.1)。
    所有业务接口需携带 Authorization: Bearer(register/login 与探针豁免);写操作需携带
    Idempotency-Key。
    限流:写接口 60 req/min/user、IP 5 req/s,超限返回 429 + Retry-After(详见 structure-contract.md 1.6)。
    5xx(INTERNAL_ERROR / GENERATION_FAILED)由全局中间件统一处理,未逐操作声明。
    列表接口 MVP 暂不分页;幂等重放返回首次成功结果(2xx,含 204)。
    旧 /pdfs 兼容路径保留过渡期,内部委托项目接口同一业务语义(structure-contract 6.2)。
servers:
  - url: https://shanka.kbzz1.top
security:
  - BearerAuth: []
...
paths:
...
  /projects/{project_id}/tasks:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [tasks]
      summary: 建立 DRAFT 任务(自动保存语义;保存章节、目标牌组和配置)
      operationId: createTask
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: 任务已创建(DRAFT)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '429':
          $ref: '#/components/responses/RateLimited'
  /tasks:
    get:
      tags: [tasks]
      summary: 学习页任务区与历史列表
      operationId: listTasks
      parameters:
        - name: project_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/TaskStatus'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /tasks/{task_id}:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    get:
      tags: [tasks]
      summary: 任务详情(七态、internal_stage、样卡、失败码)
      operationId: getTask
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [tasks]
      summary: 仅 DRAFT/AWAITING_SAMPLE_CONFIRMATION 可改配置;修改后样卡失效
      operationId: updateTask
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateRequest'
      responses:
        '200':
          description: 更新后的任务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
    delete:
      tags: [tasks]
      summary: 删除终态任务(delete_generated_cards 选择是否删除其已发布卡)
      operationId: deleteTask
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: delete_generated_cards
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: 已删除
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
  /tasks/{task_id}/samples:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    post:
      tags: [tasks]
      summary: 持久化生成 1~3 张样卡(只为比例大于 0 的难度各生成 1 张;幂等键防重复触发)
      operationId: generateTaskSamples
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          description: 上游校验不可用(API_KEY_UNAVAILABLE),可重试
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tasks/{task_id}/start:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    post:
      tags: [tasks]
      summary: 校验样卡 hash 后进入 GENERATING(过期样卡返回 409 SAMPLE_STALE)
      operationId: startTask
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: 已进入生成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '429':
          $ref: '#/components/responses/RateLimited'
  /tasks/{task_id}/abandon:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    post:
      tags: [tasks]
      summary: 放弃任务(只允许正式生成前状态,进入 ABANDONED 终态)
      operationId: abandonTask
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: 已放弃
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
  /tasks/{task_id}/retry:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    post:
      tags: [tasks]
      summary: 失败任务创建关联新任务(可沿用已确认样卡)
      operationId: retryTask
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: 新任务已创建(DRAFT,retry_of_task_id 指向原任务)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ---------- 牌组与卡片 ----------
...
  /tasks/{task_id}/batches:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    get:
      tags: [tasks]
      summary: 批次列表,含单卡 Rubric 汇总、质量分布与 Prompt Cache 记录(AC-07 观测,契约 6.9)
      operationId: listTaskBatches
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Batch'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------- 数据看板(V2.5) ----------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: opaque Bearer session token(注册/登录签发,服务端只存 SHA-256 摘要;决策 D-05)
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: 幂等键,客户端生成;重复请求返回首次成功结果(见 structure-contract.md 1.3)
...
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    TaskStatus:
      type: string
      enum: [DRAFT, SAMPLE_GENERATING, AWAITING_SAMPLE_CONFIRMATION, GENERATING, COMPLETED, FAILED, ABANDONED]
      description: V2.5 七态(structure-contract 4.1);历史 PENDING→DRAFT、RUNNING→GENERATING、CANCELLED→ABANDONED、PAUSED→FAILED(LEGACY_PAUSED_TASK)
    TaskInternalStage:
      type: string
      enum: [PLANNING, GENERATING, SCORING, PUBLISHING]
      description: 运行期内部阶段观测;不直接作为用户状态
    FailureStage:
      type: string
      enum: [PLANNING, GENERATING, SCORING, PUBLISHING]
    CoverageMode:
      type: string
      enum: [COMPACT, BALANCED, EXTENSIVE]
      description: V2.5 覆盖深度(原 quantity_tendency):语义覆盖范围,不承诺数量
...
    BatchStatus:
      type: string
      enum: [PENDING, PROCESSING, SUCCEEDED, FAILED, SKIPPED]
...
    Error:
      type: object
      required: [error]
      description: 统一错误响应,外层 error 包装(契约 1.4)
      properties:
        error:
          type: object
          required: [code, message, localization_key]
          properties:
            code:
              type: string
              description: 稳定错误码(见 structure-contract.md 第 7 章)
            message:
              type: string
              description: 面向用户的补充信息,不含内部细节
            localization_key:
              type: string
            actions:
              type: [array, 'null']
              items:
                type: string
              description: 可执行动作提示,如 ABANDON_AND_RETRY / VIEW_TASKS / WAIT_FOR_TERMINAL
            details:
              type: [object, 'null']
              additionalProperties: true
...
    Chapter:
      type: object
      required: [chapter_id, name, start_page, end_page]
      properties:
        chapter_id:
          type: string
          format: uuid
        name:
          type: string
        start_page:
          type: integer
          minimum: 1
        end_page:
          type: integer
          minimum: 1
...
    DifficultyRatio:
      type: object
      required: [basic, understanding, deep_question]
      properties:
        basic:
          type: integer
          minimum: 0
          maximum: 100
          multipleOf: 10
          description: 基础记忆(10% 整数档,允许 0),默认 40
        understanding:
          type: integer
          minimum: 0
          maximum: 100
          multipleOf: 10
          description: 理解分析(10% 整数档,允许 0),默认 40
        deep_question:
          type: integer
          minimum: 0
          maximum: 100
          multipleOf: 10
          description: 开放深问(10% 整数档,允许 0),默认 20
      description: 三档合计必须为 100;比例为 0 的难度不生成单元和样卡
    GenerationConfig:
      type: object
      required: [coverage_mode, difficulty_ratio]
      properties:
        coverage_mode:
          $ref: '#/components/schemas/CoverageMode'
        difficulty_ratio:
          $ref: '#/components/schemas/DifficultyRatio'
        custom_requirements:
          type: [string, 'null']
    TaskCursor:
      type: object
      required: [completed_batch_count]
      properties:
        completed_batch_count:
          type: integer
    Task:
      type: object
      required:
        - task_id
        - project_id
        - file_id
        - deck_id
        - status
        - selected_chapters
        - generation_config
        - generated_card_count
        - skipped_planning_group_count
        - resumable
        - created_at
        - updated_at
      properties:
        task_id:
          type: string
          format: uuid
        project_id:
          type: [string, 'null']
          format: uuid
          description: 归属项目;新任务必填,null 只兼容迁移前已失去 PDF 的终态历史任务(只读,不可重试)
        file_id:
          type: [string, 'null']
          format: uuid
          description: 新任务必填;删除 PDF 后置 null(任务保留)
        deck_id:
          type: [string, 'null']
          format: uuid
          description: 新任务必填(同项目);删除牌组后置 null(任务保留)
        retry_of_task_id:
          type: [string, 'null']
          format: uuid
          description: 失败重试关联(只指向同用户失败任务)
        operation_id:
          type: [string, 'null']
          format: uuid
          description: 持久化生成操作标识;用于跨请求/重启后的任务对账
        status:
          $ref: '#/components/schemas/TaskStatus'
        internal_stage:
          $ref: '#/components/schemas/TaskInternalStage'
          description: 运行期内部阶段观测,不直接作为用户状态
        selected_chapters:
          type: array
          items:
            $ref: '#/components/schemas/Chapter'
          description: 开始正式生成前冻结快照(契约 4.1)
        generation_config:
          $ref: '#/components/schemas/GenerationConfig'
        sample_cards:
          type: [array, 'null']
          items:
            $ref: '#/components/schemas/SampleCard'
          description: 持久化 1~3 张样卡;配置变化时清空
        sample_config_hash:
          type: [string, 'null']
        sample_confirmed_at:
          type: [string, 'null']
          format: date-time
        cursor:
          $ref: '#/components/schemas/TaskCursor'
        generated_card_count:
          type: integer
          description: 只统计已发布卡;失败任务为 0
        total_batch_count:
          type: [integer, 'null']
        completed_batch_count:
          type: [integer, 'null']
        completion_reason:
          type: [string, 'null']
          description: NO_GENERATION_UNITS(全组规划成功但 0 个合法单元的业务空结果,COMPLETED)
        skipped_planning_group_count:
          type: integer
          description: 部分规划组失败被跳过的组数(仅部分成功时 > 0)
        resumable:
          type: boolean
          description: 内部租约恢复判定(用户侧无 resume API)
        failure_stage:
          $ref: '#/components/schemas/FailureStage'
        error_code:
          type: [string, 'null']
        created_at:
          type: string
          format: date-time
        started_at:
          type: [string, 'null']
          format: date-time
        ended_at:
          type: [string, 'null']
          format: date-time
        updated_at:
          type: string
          format: date-time
    TaskCreateRequest:
      type: object
      required: [deck_id, chapter_ids, generation_config]
      properties:
        deck_id:
          type: string
          format: uuid
          description: 目标牌组(必须属于同一项目)
        chapter_ids:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
        generation_config:
          $ref: '#/components/schemas/GenerationConfig'
      description: 建立 DRAFT 任务(project_id 取自路径);请求中不携带 API Key,服务端使用已保存 Key
    TaskUpdateRequest:
      type: object
      properties:
        deck_id:
          type: string
          format: uuid
          description: 目标牌组(必须属于同一项目)
        chapter_ids:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
        generation_config:
          $ref: '#/components/schemas/GenerationConfig'
      description: 仅 DRAFT/AWAITING_SAMPLE_CONFIRMATION 可改;修改后样卡失效(sample_cards/sample_config_hash/sample_confirmed_at 清空)
    SampleCard:
      type: object
      required: [card_id, front, back, card_type]
      properties:
        card_id:
          type: string
          format: uuid
        front:
          type: string
        back:
          type: string
        code:
          type: [string, 'null']
          description: 卡片编号
        card_type:
          type: string
          enum: [QUESTION, TRUE_FALSE]
        question:
          type: [string, 'null']
          description: 仅 QUESTION 卡
        answer:
          type: [string, 'null']
          description: 仅 QUESTION 卡
        statement:
          type: [string, 'null']
          description: 仅 TRUE_FALSE 卡
        answer_boolean:
          type: [boolean, 'null']
          description: 仅 TRUE_FALSE 卡
        explanation:
          type: [string, 'null']
          description: 仅 TRUE_FALSE 卡
        target_difficulty:
          type: [string, 'null']
          enum: [BASIC, UNDERSTANDING, DEEP_QUESTION]
          description: 生成时目标难度
...
    Batch:
      type: object
      description: 内部资源(契约 3.7;批=生成单元,每单元 1 批 1 次生成调用,每单元 1 卡);经 GET /v1/tasks/{task_id}/batches 观测返回(契约 6.9)。generation_unit_id 为内部列(观测归因用),视图不返回;尝试数与 token 权威在 llm_call_attempts 账本,Batch token/版本列为生成阶段兼容投影
      required: [batch_id, task_id, batch_index, status, retry_count]
      properties:
        batch_id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        batch_index:
          type: integer
        status:
          $ref: '#/components/schemas/BatchStatus'
        generated_item_ids:
          type: [array, 'null']
          items:
            type: string
        retry_count:
          type: integer
        coverage_rate:
          type: [number, 'null']
        duplicate_rate:
          type: [number, 'null']
        difficulty_distribution:
          type: [object, 'null']
        chapter_distribution:
          type: [object, 'null']
        card_type_distribution:
          type: [object, 'null']
        cache_hit_tokens:
          type: [integer, 'null']
        cache_miss_tokens:
          type: [integer, 'null']
        output_tokens:
          type: [integer, 'null']
        difficulty_deviation:
          type: [number, 'null']
        request_id:
          type: [string, 'null']
        model:
          type: [string, 'null']
        prompt_version:
          type: [string, 'null']
        schema_version:
          type: [string, 'null']
        rubric_version:
          type: [string, 'null']
        duration_ms:
          type: [integer, 'null']
        http_status:
          type: [integer, 'null']
        cost_estimate:
          type: [number, 'null']
          description: 估算金额(元, 按生效日期价格常量换算, 仅观测不落库; 契约 8.4)
        created_at:
          type: string
          format: date-time
        ended_at:
          type: [string, 'null']
          format: date-time
...
  responses:
    BadRequest:
      description: 请求非法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: >-
        受保护接口:缺失、非法、撤销或过期的 Bearer 会话(AUTH_REQUIRED / AUTH_INVALID,携带 WWW-Authenticate: Bearer);
        登录接口:凭据错误(INVALID_CREDENTIALS)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: 状态冲突或幂等内容不一致
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unprocessable:
      description: 业务校验失败(如 PDF_TOC_MISSING、API_KEY_INVALID)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: 限流
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'